moved productivity script to outside file
